Moved css files into a seperate folder "src/styles" and updated the respective paths in pages
