+++ frontend/src/components/SearchBox.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect, useRef } from 'react';
 import { useNavigate } from 'react-router-dom';
-import './SearchBox.css';
+import '../styles/SearchBox.css';
 
 const SearchBox = () => {
     const [searchQuery, setSearchQuery] = useState('');
