+++ frontend/src/components/Sidebar.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import './Sidebar.css';
+import '../styles/Sidebar.css';
 import { useNavigate } from 'react-router-dom';
 
 const Sidebar = ({ userProfile }) => {
