+++ frontend/src/styles/NotificationList.css
