+++ frontend/src/styles/Post.css
