+++ frontend/src/styles/PostForm.css
