+++ frontend/src/styles/SearchBox.css
