+++ frontend/src/styles/Sidebar.css
