+++ frontend/src/styles/ViewPosts.css
